finished model loading test (1/2)
This is all copy-pasted. I'm not that proud of me (or even proud of me at all), but there was so many bugs I couldn't handle it. I finnaly figured out why i had a bug with linking and how to solve this, but I already did the thing.... so